+++ tests/integration/prerender.spec.ts
@@ -2,18 +2,19 @@ import { test, expect, Page, Browser } from "@playwright/test";
 import { chromium } from "playwright";
 import Flags from "../../src/flags.js";
 import BeaconStore from "../helpers/beacon-store.js";
-import { getNavTiming, hasFlag } from "../helpers/lux.js";
+import {
+  getCpuStat,
+  getNavTiming,
+  getSearchParam,
+  hasFlag,
+  parseUserTiming,
+} from "../helpers/lux.js";
 
 test.describe("LUX prerender support", () => {
   test.skip(
     ({ browserName }) => browserName !== "chromium",
     "Prerendering is only supported in Chromium"
   );
-
-  // The tests in this file rely on the SQLite beacon store to inspect beacon requests. This is
-  // because network requests from the prerender process cannot be intercepted. To make sure that
-  // requests from each test do not interfere with each other, we run these tests in serial.
-  test.describe.configure({ mode: "serial" });
 
   let browser: Browser, page: Page, store: BeaconStore;
 
@@ -30,23 +31,26 @@ test.describe("LUX prerender support", () => {
   });
 
   test.beforeEach(async () => {
-    await store.deleteByPathname("/prerender-index.html");
-    await store.deleteByPathname("/prerender-page.html");
+    await store.deleteAll();
   });
 
   test("pages loaded by prerender speculation rules do not trigger beacons", async () => {
-    await page.goto("/prerender-index.html?useBeaconStore", { waitUntil: "networkidle" });
+    await page.goto(`/prerender-index.html?useBeaconStore=${store.id}`, {
+      waitUntil: "networkidle",
+    });
 
     const beacons = await store.findAll();
     expect(beacons.length).toEqual(1);
     expect(beacons[0].pathname).toEqual("/prerender-index.html");
   });
 
-  test("LUX.autoWhenHidden=true sends the beacon on prerendered pages", async () => {
-    await page.goto("/prerender-index.html?useBeaconStore&injectScript=LUX.autoWhenHidden=true");
+  test("LUX.trackHiddenPages=true sends the beacon on prerendered pages", async () => {
+    await page.goto(
+      `/prerender-index.html?useBeaconStore=${store.id}&injectScript=LUX.trackHiddenPages=true`
+    );
 
     // Wait for up to 5 seconds for there to be 2 beacons in the beacon store
-    await expect.poll(async () => (await store.findAll()).length, { timeout: 5000 }).toEqual(2);
+    await expect.poll(() => store.countAll(), { timeout: 5000 }).toEqual(2);
 
     const beacons = await store.findAll();
 
@@ -59,16 +63,19 @@ test.describe("LUX prerender support", () => {
     expect(hasFlag(new URL(beacons[1].url), Flags.VisibilityStateNotVisible)).toBe(true);
   });
 
-  test("Prerendered pages a non-zero value for activationStart", async () => {
+  test("metrics for prerendered pages are relative to activationStart", async () => {
+    const LONG_TASK_TIME = 70;
     const CLICK_WAIT_TIME = 500;
     const IMAGE_DELAY_TIME = 1000;
-    const SEND_LUX_DELAY = CLICK_WAIT_TIME + IMAGE_DELAY_TIME + 100;
 
     await page.goto(
       [
-        "/prerender-index.html?useBeaconStore",
+        `/prerender-index.html?useBeaconStore=${store.id}`,
         `imageDelay=${IMAGE_DELAY_TIME}`,
-        `injectScript=LUX.auto=false;setTimeout(LUX.send, ${SEND_LUX_DELAY})`,
+        `injectScript=${[
+          `LUX.minMeasureTime=${IMAGE_DELAY_TIME + CLICK_WAIT_TIME}`,
+          `if (location.pathname === "/prerender-page.html") { createLongTask(${LONG_TASK_TIME}); }`,
+        ].join(";")}`,
       ].join("&")
     );
 
@@ -108,6 +115,36 @@ test.describe("LUX prerender support", () => {
     expect(getNavTiming(beacon, "oc")).toEqual(0);
     expect(getNavTiming(beacon, "ls")).toEqual(0);
     expect(getNavTiming(beacon, "le")).toEqual(0);
+
+    // Element timing
+    const ET = parseUserTiming(getSearchParam(beacon, "ET"));
+
+    // The first image should have loaded before activationStart
+    expect(ET["eve-image"].startTime).toBeLessThan(activationStart);
+
+    // The second image was delayed and should have loaded after IMAGE_DELAY TIME, but relative
+    // to activationStart.
+    expect(ET["charlie-image"].startTime).toBeGreaterThanOrEqual(
+      IMAGE_DELAY_TIME - activationStart
+    );
+
+    // Paint metrics
+    expect(getNavTiming(beacon, "sr")).toBeLessThan(activationStart);
+    expect(getNavTiming(beacon, "fc")).toBeLessThan(activationStart);
+    expect(getNavTiming(beacon, "lc")).toBeGreaterThanOrEqual(ET["charlie-image"].startTime);
+
+    // CPU metrics
+    const longTaskCount = getCpuStat(beacon, "n");
+    const longTaskTotal = getCpuStat(beacon, "s");
+    const longTaskMedian = getCpuStat(beacon, "d");
+    const longTaskMax = getCpuStat(beacon, "x");
+    const firstCpuIdle = getCpuStat(beacon, "i");
+
+    expect(longTaskCount).toEqual(1);
+    expect(longTaskTotal).toBeGreaterThanOrEqual(LONG_TASK_TIME);
+    expect(longTaskMedian).toEqual(longTaskTotal);
+    expect(longTaskMax).toEqual(longTaskTotal);
+    expect(firstCpuIdle).toBeLessThan(activationStart);
 
     // The same Playwright bug mentioned above means we have to forcefully close the page after this test
     await page.close();